# [sr][pyper] add fusion broadcast_concat_batch_matmul_batch_gather